\name{download.data.dictionary}
\alias{download.data.dictionary}
\title{
Download UKB Data Dictionary
}
\description{
Download UKB Data Dictionary to user specified folder.
}
\usage{
download.data.dictionary(output_path)
}

\arguments{
  \item{output_path}{
  The output folder where the UKB Data Dictionary file will be downloaded.
}
}

\value{
If successful, the current UKB Data Dictionary will be downloaded to the folder sepcificed by user.
}


\examples{
# Download the UKB Data Dictionary to Current working directory
# download.data.dictionary(getwd())
}
\keyword{ UKB Data Dictionary }

