\name{is.tree.code}
\alias{is.tree.code}
\title{
  Check if the input UKB coding file has a tree structure.
}
\description{
Check if the input UKB coding file has a tree structure. Return TRUE if the coding has a tree structure and FALSE if not.
}
\usage{
is.tree.code(data_code_file)
}
\arguments{
  \item{data_code_file}{ The file path/name of a UKB data coding file.}
}

\value{
    Return TRUE if the coding has a tree structure and FALSE if not.
}
