\name{PEACOK-package}
\alias{PEACOK-package}
\alias{PEACOK}
\docType{package}
\title{
\packageTitle{PEACOK}
}
\description{
\packageDescription{PEACOK}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PEACOK}
\packageIndices{PEACOK}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{PEACOK}

Maintainer: \packageMaintainer{PEACOK}
}
\references{
Literature or other references for background information 
}
\keyword{ package }
\seealso{
To be done
}
\examples{
#example here 
}
