\name{read.pheno.matrix}
\alias{read.pheno.matrix}
\title{
Read the whole or a trunk of the Phenotype Matrix by columns. 
}
\description{
Read the whole or a trunk of the Phenotype Matrix by columns. 
}
\usage{
read.pheno.matrix(matrix_file, pheno_start = NULL, pheno_end = NULL)
}
\arguments{
  \item{matrix_file}{The Phenotype matrix file to be read. }
  \item{pheno_start}{The start index of the phenotype columns to be read. }
  \item{pheno_end}{The end index of the phenotype columns to be read. 
}
}
\details{
Returns a data frame that holds the whole or a trunk of the phenotype matrix.
}

