\name{run.categorical.ordered}
\alias{run.categorical.ordered}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Run association test for Categorical Ordered phenotypes
}
\description{
Run association test for "Categorical Ordered" phenotypes.
}
\usage{
run.categorical.ordered(pheno, geno, confounders)
}
\arguments{
  \item{pheno}{ the phenotype vector}
  \item{geno}{ the genotype vector}
  \item{confounders}{the confounder matrix}
}

\value{
Return a list of test statistics and summaries
}