\name{run.continuous}
\alias{run.continuous}
\title{
Run association test for continuous phenotypes
}
\description{
Run association test for continuous phenotypes.
}
\usage{
run.continuous(pheno,geno,confounders)
}
\arguments{
  \item{pheno}{ the phenotype vector}
  \item{geno}{ the genotype vector}
  \item{confounders}{the confounder matrix}
}

\value{
Return a list of test statistics and summaries
}
